Added SupplySchedule interface

+++ contracts/libraries/Math.sol
@@ -6,8 +6,8 @@ pragma solidity ^0.8.0;
 import "./SafeDecimalMath.sol";
 
 library Math {
-    using SafeMath for uint;
-    using SafeDecimalMath for uint;
+    using SafeMath for uint256;
+    using SafeDecimalMath for uint256;
 
     /**
      * @dev Uses "exponentiation by squaring" algorithm where cost is 0(logN)
@@ -15,10 +15,10 @@ library Math {
      * Calculates x^n with x as fixed-point and n as regular unsigned int.
      * Calculates to 18 digits of precision with SafeDecimalMath.unit()
      */
-    function powDecimal(uint x, uint n) internal pure returns (uint) {
+    function powDecimal(uint256 x, uint256 n) internal pure returns (uint256) {
         // https://mpark.github.io/programming/2014/08/18/exponentiation-by-squaring/
 
-        uint result = SafeDecimalMath.unit();
+        uint256 result = SafeDecimalMath.unit();
         while (n > 0) {
             if (n % 2 != 0) {
                 result = result.multiplyDecimal(x);
